Run bootstrapping of Spearman correlations within a dataset

This function provides the more low-level functionality of bootstrapping the Spearman correlations of the columns within a dataset. Only use this function if you want a low-level interface, else run_csd provides a more streamlined approach if you want to do a CSD analysis.

run_cor_bootstrap(
  x,
  n_it = 20L,
  nThreads = 1L,
  verbose = TRUE,
  iterations_gap = 1L
)

Arguments

x

Numeric matrix, the gene expression matrix to analyse. Genes are in columns, samples are in rows.

n_it

Integer, number of bootstrap iterations

nThreads

Integer, number of threads to use for computations

verbose

Logical, should progress be printed?

iterations_gap

If output is verbose - number of iterations after issue a status message (Default=1 - displayed only if verbose=TRUE)

Value

A list with two fields

rho

Numeric matrix constaining the bootstrapped mean of the Spearman correlation between each column

var

Numeric matrix constaining the bootstrapped variance of the Spearman correlation between each column

Examples

data("normal_expression")
cor_res <- run_cor_bootstrap(
    x = normal_expression,
    n_it = 100, nThreads = 2L
)
#> Fri Mar  4 14:13:29 2022 => Running bootstrap iteration 1 of 100...
#> Fri Mar  4 14:13:29 2022 => Running bootstrap iteration 2 of 100...
#> Fri Mar  4 14:13:29 2022 => Running bootstrap iteration 3 of 100...
#> Fri Mar  4 14:13:29 2022 => Running bootstrap iteration 4 of 100...
#> Fri Mar  4 14:13:30 2022 => Running bootstrap iteration 5 of 100...
#> Fri Mar  4 14:13:30 2022 => Running bootstrap iteration 6 of 100...
#> Fri Mar  4 14:13:30 2022 => Running bootstrap iteration 7 of 100...
#> Fri Mar  4 14:13:30 2022 => Running bootstrap iteration 8 of 100...
#> Fri Mar  4 14:13:31 2022 => Running bootstrap iteration 9 of 100...
#> Fri Mar  4 14:13:31 2022 => Running bootstrap iteration 10 of 100...
#> Fri Mar  4 14:13:31 2022 => Running bootstrap iteration 11 of 100...
#> Fri Mar  4 14:13:32 2022 => Running bootstrap iteration 12 of 100...
#> Fri Mar  4 14:13:32 2022 => Running bootstrap iteration 13 of 100...
#> Fri Mar  4 14:13:32 2022 => Running bootstrap iteration 14 of 100...
#> Fri Mar  4 14:13:32 2022 => Running bootstrap iteration 15 of 100...
#> Fri Mar  4 14:13:32 2022 => Running bootstrap iteration 16 of 100...
#> Fri Mar  4 14:13:33 2022 => Running bootstrap iteration 17 of 100...
#> Fri Mar  4 14:13:33 2022 => Running bootstrap iteration 18 of 100...
#> Fri Mar  4 14:13:33 2022 => Running bootstrap iteration 19 of 100...
#> Fri Mar  4 14:13:33 2022 => Running bootstrap iteration 20 of 100...
#> Fri Mar  4 14:13:33 2022 => Running bootstrap iteration 21 of 100...
#> Fri Mar  4 14:13:34 2022 => Running bootstrap iteration 22 of 100...
#> Fri Mar  4 14:13:34 2022 => Running bootstrap iteration 23 of 100...
#> Fri Mar  4 14:13:34 2022 => Running bootstrap iteration 24 of 100...
#> Fri Mar  4 14:13:34 2022 => Running bootstrap iteration 25 of 100...
#> Fri Mar  4 14:13:34 2022 => Running bootstrap iteration 26 of 100...
#> Fri Mar  4 14:13:35 2022 => Running bootstrap iteration 27 of 100...
#> Fri Mar  4 14:13:35 2022 => Running bootstrap iteration 28 of 100...
#> Fri Mar  4 14:13:35 2022 => Running bootstrap iteration 29 of 100...
#> Fri Mar  4 14:13:35 2022 => Running bootstrap iteration 30 of 100...
#> Fri Mar  4 14:13:35 2022 => Running bootstrap iteration 31 of 100...
#> Fri Mar  4 14:13:36 2022 => Running bootstrap iteration 32 of 100...
#> Fri Mar  4 14:13:36 2022 => Running bootstrap iteration 33 of 100...
#> Fri Mar  4 14:13:36 2022 => Running bootstrap iteration 34 of 100...
#> Fri Mar  4 14:13:36 2022 => Running bootstrap iteration 35 of 100...
#> Fri Mar  4 14:13:36 2022 => Running bootstrap iteration 36 of 100...
#> Fri Mar  4 14:13:37 2022 => Running bootstrap iteration 37 of 100...
#> Fri Mar  4 14:13:37 2022 => Running bootstrap iteration 38 of 100...
#> Fri Mar  4 14:13:37 2022 => Running bootstrap iteration 39 of 100...
#> Fri Mar  4 14:13:37 2022 => Running bootstrap iteration 40 of 100...
#> Fri Mar  4 14:13:38 2022 => Running bootstrap iteration 41 of 100...
#> Fri Mar  4 14:13:38 2022 => Running bootstrap iteration 42 of 100...
#> Fri Mar  4 14:13:38 2022 => Running bootstrap iteration 43 of 100...
#> Fri Mar  4 14:13:38 2022 => Running bootstrap iteration 44 of 100...
#> Fri Mar  4 14:13:39 2022 => Running bootstrap iteration 45 of 100...
#> Fri Mar  4 14:13:39 2022 => Running bootstrap iteration 46 of 100...
#> Fri Mar  4 14:13:39 2022 => Running bootstrap iteration 47 of 100...
#> Fri Mar  4 14:13:39 2022 => Running bootstrap iteration 48 of 100...
#> Fri Mar  4 14:13:39 2022 => Running bootstrap iteration 49 of 100...
#> Fri Mar  4 14:13:40 2022 => Running bootstrap iteration 50 of 100...
#> Fri Mar  4 14:13:40 2022 => Running bootstrap iteration 51 of 100...
#> Fri Mar  4 14:13:40 2022 => Running bootstrap iteration 52 of 100...
#> Fri Mar  4 14:13:40 2022 => Running bootstrap iteration 53 of 100...
#> Fri Mar  4 14:13:40 2022 => Running bootstrap iteration 54 of 100...
#> Fri Mar  4 14:13:41 2022 => Running bootstrap iteration 55 of 100...
#> Fri Mar  4 14:13:41 2022 => Running bootstrap iteration 56 of 100...
#> Fri Mar  4 14:13:41 2022 => Running bootstrap iteration 57 of 100...
#> Fri Mar  4 14:13:41 2022 => Running bootstrap iteration 58 of 100...
#> Fri Mar  4 14:13:41 2022 => Running bootstrap iteration 59 of 100...
#> Fri Mar  4 14:13:42 2022 => Running bootstrap iteration 60 of 100...
#> Fri Mar  4 14:13:42 2022 => Running bootstrap iteration 61 of 100...
#> Fri Mar  4 14:13:42 2022 => Running bootstrap iteration 62 of 100...
#> Fri Mar  4 14:13:42 2022 => Running bootstrap iteration 63 of 100...
#> Fri Mar  4 14:13:42 2022 => Running bootstrap iteration 64 of 100...
#> Fri Mar  4 14:13:43 2022 => Running bootstrap iteration 65 of 100...
#> Fri Mar  4 14:13:43 2022 => Running bootstrap iteration 66 of 100...
#> Fri Mar  4 14:13:43 2022 => Running bootstrap iteration 67 of 100...
#> Fri Mar  4 14:13:43 2022 => Running bootstrap iteration 68 of 100...
#> Fri Mar  4 14:13:43 2022 => Running bootstrap iteration 69 of 100...
#> Fri Mar  4 14:13:44 2022 => Running bootstrap iteration 70 of 100...
#> Fri Mar  4 14:13:44 2022 => Running bootstrap iteration 71 of 100...
#> Fri Mar  4 14:13:44 2022 => Running bootstrap iteration 72 of 100...
#> Fri Mar  4 14:13:44 2022 => Running bootstrap iteration 73 of 100...
#> Fri Mar  4 14:13:44 2022 => Running bootstrap iteration 74 of 100...
#> Fri Mar  4 14:13:45 2022 => Running bootstrap iteration 75 of 100...
#> Fri Mar  4 14:13:45 2022 => Running bootstrap iteration 76 of 100...
#> Fri Mar  4 14:13:45 2022 => Running bootstrap iteration 77 of 100...
#> Fri Mar  4 14:13:45 2022 => Running bootstrap iteration 78 of 100...
#> Fri Mar  4 14:13:45 2022 => Running bootstrap iteration 79 of 100...
#> Fri Mar  4 14:13:46 2022 => Running bootstrap iteration 80 of 100...
#> Fri Mar  4 14:13:46 2022 => Running bootstrap iteration 81 of 100...
#> Fri Mar  4 14:13:46 2022 => Running bootstrap iteration 82 of 100...
#> Fri Mar  4 14:13:46 2022 => Running bootstrap iteration 83 of 100...
#> Fri Mar  4 14:13:46 2022 => Running bootstrap iteration 84 of 100...
#> Fri Mar  4 14:13:47 2022 => Running bootstrap iteration 85 of 100...
#> Fri Mar  4 14:13:47 2022 => Running bootstrap iteration 86 of 100...
#> Fri Mar  4 14:13:47 2022 => Running bootstrap iteration 87 of 100...
#> Fri Mar  4 14:13:47 2022 => Running bootstrap iteration 88 of 100...
#> Fri Mar  4 14:13:47 2022 => Running bootstrap iteration 89 of 100...
#> Fri Mar  4 14:13:48 2022 => Running bootstrap iteration 90 of 100...
#> Fri Mar  4 14:13:48 2022 => Running bootstrap iteration 91 of 100...
#> Fri Mar  4 14:13:48 2022 => Running bootstrap iteration 92 of 100...
#> Fri Mar  4 14:13:48 2022 => Running bootstrap iteration 93 of 100...
#> Fri Mar  4 14:13:48 2022 => Running bootstrap iteration 94 of 100...
#> Fri Mar  4 14:13:49 2022 => Running bootstrap iteration 95 of 100...
#> Fri Mar  4 14:13:49 2022 => Running bootstrap iteration 96 of 100...
#> Fri Mar  4 14:13:49 2022 => Running bootstrap iteration 97 of 100...
#> Fri Mar  4 14:13:49 2022 => Running bootstrap iteration 98 of 100...
#> Fri Mar  4 14:13:49 2022 => Running bootstrap iteration 99 of 100...
#> Fri Mar  4 14:13:50 2022 => Running bootstrap iteration 100 of 100...